folder structure
https://towardsdatascience.com/whats-init-for-me-d70a312da583
