+++ setup.py
@@ -1,0 +1,18 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Thu Feb  6 15:00:52 2020
+
+@author: OediP
+"""
+
+from setuptools import setup
+
+setup(name='segmentation_tools',
+      version='0.1',
+      description='The funniest joke in the world',
+      url='http://github.com/philipoedi/segmentation_tools',
+      author='Flying Circus',
+      author_email='flyingcircus@example.com',
+      license='MIT',
+      packages=['segmentation_tools'],
+      zip_safe=False)
